ScannerAI-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pillow>=10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0


